Added exception handling to the "HunspellDictionary" class
As the current implementation doesn't really support anything other than the Danish dictionary, this commit adds a try/catch clause to catch if loading the dictionary fails. If this is the case, an instance of "HunspellDictionary" will still be returned, but with "Exception" property set to the exception from the load operation, and with the "IsSuccessful" property set to "false".