Misc changes to parsing dictionary files
Since different dictionary files use a few different formats for referencing the affix rule set, this commit tries to add better support for this. Some dictionaries specify the flags as a comma separated value with numeric IDs, while others specify a number of non-separated letters where each letter is a reference to a rule set in the affix file.